Refactor collaboration page to handle smaller windows and also move logic of changing languages to parent component of CollabView.
