+++ client/src/components/userCollaboration/viewComponents/collabView.css
@@ -1,17 +1,19 @@
-.question-details-component {
+.collab-view {
     display: flex;
     height: 100vh;
+    max-width: 100vw;
     box-sizing: border-box;
-    padding: 20px 20px 15px 20px;
+    padding: 20px;
     background-color: lightgray;
+    column-gap: 10px;
 }
 
 .question-details {
-    flex: 1 0;
+    flex: 1 1;
     background-color: white;
-    margin: 0 10px;
     border-radius: 8px;
     padding: 15px 20px;
+    position: relative;
 }
 
 .question-title {
@@ -24,9 +26,10 @@
 }
 
 .return-icon {
-    font-size: 40px;
+    font-size: 35px;
     cursor: pointer;
     border-radius: 10px;
-    bottom: 20px;
+    top: 20px;
+    right: 20px;
     position: absolute;
 }
